Remove "magit-libgit" and "libgit" recipes
I should never have added "magit-libgit" to Melpa, as it is only a
proof-of-concept.  "Using it" would currently involve contributing
to it, which cannot be done through Melpa.

"libgit" is one of only two packages that still depend on support
for submodules in "package-build.el"; a feature which is about to
be removed.  No other packages depend on it, but some users might
use it in their personal code or an unpublished package.  I ended
up maintaining this package and intend to bring it back to Melpa,
but how soon I'll do that might depend on whether anyone informs
me that that would benefit them.

See https://github.com/melpa/package-build/issues/62.