Remove "journalctl-mode" package due to licensing issues
It is released under GPL-2.0-only, which is not compatible with the
license used by Emacs.  I don't think that was the author's intention
but legally speaking that is where we stand.  I was unable to get the
author's attention to get this sorted out, leaving me no choice but to
remove this package.

See https://github.com/SebastianMeisel/journalctl-mode/issues/18.