Add recipe for dired-rsync-transient
This is a useful dired-rsync command but I packaged it as separate
file to avoid the extra dependency for the base package. I've also
cleaned up the formatting for the dired-rsync recipe and set an
explicit file for it as well to avoid pulling in the testing files
from the repo.

See: https://github.com/stsquad/dired-rsync/issues/40