Fix the updates template so it works for MELPA Stable.
We detect if we're using melpa stable by looking at the path for the script.
This is a bit hacky but a lot of the build stuff is hacky surrounding the
stable repo. This should be fine and is better than forking the file and having
to maintain two different versions.