Remove "midje-mode" package
The upstream repository has been removed, after not being touched in
years.  The package is also in a state that would prevent it from
being added to Melpa as-is.  If someone still uses this and would like
for it to be added back to Melpa, they will have to put in some work.