make: Report which directories are being cleaned
Depending on the value of $STABLE different directories are being
cleaned out, so one has to use "make clean; make STABLE=t clean"
to clean everything, making it desirable if these targets report
which directories are being cleaned.

"clean" should probably clean all directories, but how that should
be implemented depends on upcoming changes, so we don't do that for
now.