make: Use $EMACS instead of $EMACS_COMMAND
This effectively reverts [1: 52757abf] from 2014.

Since 2006, INSIDE_EMACS is set to t in subshells running inside Emacs
but for backward compatibility EMACS was still also set to indicate
that the process was running inside Emacs.  Since [2: beaab898968]
from 2015, and first released as part of Emacs 25.1, EMACS is not set
for that purpose anymore.

By now it is customary in Makefiles to use the EMACS variable to allow
selecting a specific Emacs executable.

1: 2014-07-11 52757abf2ed346e01833d03007f5df313630cd2b
   Use $EMACS_COMMAND rather than just $EMACS (see #1855)

2: 2006-11-18 142260f79d989ccbdc7bcf7e14dea2602aab931c
   Addition of INSIDE_EMACS, and deprecation of EMACS.

2: 2015-04-09 beaab898968caf8b243a33d24824d430fabc31fc
   Stop messing with the EMACS env var