make: Consistently use " to quote strings
> You should never quote anything because of make. Make doesn't
> understand or parse single- or double-quote characters in any
> way. Every quoting character you write in a makefile will be
> kept as a literal quote and passed along as-is to the commands
> that make invokes.

  -- MadScientist (https://stackoverflow.com/questions/23330243)

With that in mind, let's have a look at what the most "complex"
strings are in our Makefile: s-expressions that contain both
quoted symbols and strings.  I.e., we have to use " for quoting
the string for the shell.

So let's just do that for all strings.  That way, we will do it
right, even if we forget that make doesn't deal with single- and
double-quote characters itself; provided of course we continue
to try to be consistent.