Fix geiser-stklos :files property
It seems that the property has never had the correct value.

The initial recipe had

    :files (:defaults ("geiser.stk"))

which is correct, except that the parentheses around "geiser.stk" make
that file unrecognized. The property was interpreted as equal to

    :files (:defaults)

which caused it to be removed in the batch maintenance commit
ad320d60e2c95881f31628c19ad3b9ece7e3d165.

The present commit adds back "geiser.stk" without the parentheses.