Command Log
tl;dr The package this forks, command-log-mode, is pretty bad.  I did some work
to update it and never added it to MELPA.  The feature and recipe name for the
original are also kind of bad.  I would just replace it, but that might brake
people's configs.  The command names changed, and I didn't alias the originals.

The original command-log-mode did not properly declare a minor mode and had a
bunch of other issues.  This version incorporated the open PR's and then
attempted to get to a healthier level.

However, I don't intend to expand this package because after doing this work, I
found keycast.  At some point I confused keycast.el with some Mac program that I
obviously couldn't use on Linux.  I added a link or mention of Keycast in the
README and package description before writing this.

The original package still shows up in google results, and it's kind of a
garbage package, but I changed the feature name, so unless we also get rid of or
re-designate the old command-log-mode from MELPA, they are independent packages.

At the least, I'll respond to PR's which I don't expect many of.