Update two recipes to rename files more explicitly
Previously a file named "*.el.in" was renamed to "*.el".  It was still
necessary to add that file explicitly to `:files', because such files
were not matched by default files spec (to avoid false positives).

Now that files have to be renamed more explicitly, we have to adjust
these two recipes accordingly.  (These are the only two recipes that
require that a file is renamed.)