make: No longer pause after docker build
The nap is supposed to happen in between builds but when we locally
initiate a single build using the `docker-build-run' target, that
resulted in an unnecessary delay at the end of the one and only
build run.

On the server docker builds are not started using the Makefile,
so there we continue to take a five minute nap between builds.