Package: jupyter-ascending
The Jupyter Ascending package facilitates editing and executing code
in a Jupyter Python notebook from an ordinary Python buffer in
Emacs. It does this by providing Emacs commands which utilize the
Jupytext and Jupyter Ascending command line tools for
- rendering .ipynb notebooks as standard Python source files,
- synchronizing the state of the 2 files, and
- executing code in the Jupyter notebook.
This allows users to work with Jupyter Python notebooks in the comfort
of Emacs and make use of all the features they would normally have
when editing Python code, such as code completion, linting, etc.